requirements.txt
asgiref==3.7.2
autopep8==2.0.4
Django==5.0.2
django-split-settings==1.2.0
flake8==7.0.0
isort==5.13.2
mccabe==0.7.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
python-dotenv==1.0.1
split-settings==1.0.0
sqlparse==0.4.4
